pyproject.toml
[project]
name = "django-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg>=3.2.10",
    "psycopg-binary>=3.2.10",
    "python-decouple>=3.8",
]
